+++ CHANGES.md
@@ -1,3 +1,7 @@
+#2015-07-10
+
+* Changed the way output is logged: Optimisation results now appear in a log file for each case.
+
 #2015-07-09
 
 * Refactored some of go.rb